add replay window to oscore

+++ modules/oscore/src/aad.c
@@ -49,7 +49,7 @@ enum oscore_error create_aad(struct o_coap_option *options, uint16_t opt_num,
 	aad_array._aad_array_options.len = opts_i.len;
 	aad_array._aad_array_options.value = opts_i.ptr;
 
-	size_t payload_len_out;
+	uint32_t payload_len_out;
 	success_encoding = cbor_encode_aad_array(out->ptr, out->len, &aad_array,
 						 &payload_len_out);
